pyproject.toml
[tool.poetry]
name = "expensetracker"
version = "0.1.0"
description = "Expense Traker Langgraph project"
authors = ["Subash Pandey <axlesubash111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anthropic = "0.25.6"
anyio = "4.3.0"
appnope = "0.1.4"
asttokens = "2.4.1"
attrs = "23.2.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
comm = "0.2.2"
dataclasses-json = "0.6.4"
debugpy = "1.8.1"
decorator = "5.1.1"
defusedxml = "0.7.1"
distro = "1.9.0"
executing = "2.0.1"
fastapi = "0.110.2"
filelock = "3.13.4"
frozenlist = "1.4.1"
fsspec = "2024.3.1"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
huggingface-hub = "0.22.2"
idna = "3.7"
ipykernel = "6.29.4"
ipython = "8.23.0"
jedi = "0.19.1"
jsonpatch = "1.33"
jsonpointer = "2.4"
jupyter-client = "8.6.1"
jupyter-core = "5.7.2"
langchain = "0.1.16"
langchain-anthropic = "0.1.11"
langchain-community = "0.0.34"
langchain-core = "0.1.45"
langchain-openai = "0.1.3"
langchain-text-splitters = "0.0.1"
langgraph = "0.0.38"
langsmith = "0.1.49"
marshmallow = "3.21.1"
matplotlib-inline = "0.1.7"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
numpy = "1.26.4"
openai = "1.23.2"
orjson = "3.10.1"
packaging = "23.2"
parso = "0.8.4"
pexpect = "4.9.0"
platformdirs = "4.2.0"
prompt-toolkit = "3.0.43"
psutil = "5.9.8"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pydantic = "2.7.0"
pydantic-core = "2.18.1"
pygments = "2.17.2"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
pyzmq = "26.0.2"
regex = "2024.4.16"
requests = "2.31.0"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.29"
stack-data = "0.6.3"
starlette = "0.37.2"
tenacity = "8.2.3"
tiktoken = "0.6.0"
tokenizers = "0.19.1"
tornado = "6.4"
tqdm = "4.66.2"
traitlets = "5.14.3"
typing-inspect = "0.9.0"
typing-extensions = "4.11.0"
urllib3 = "2.2.1"
uvicorn = "0.29.0"
wcwidth = "0.2.13"
yarl = "1.9.4"
streamlit = "^1.37.1"
psycopg2-binary = "^2.9.9"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:main"
